Revert "feat: Adds discussions settings for new discusions experience (#28749)"
This reverts commit bb0c03123c912dd913e6d7e6c5b3a4e183d4c95b.
