+++ lms/djangoapps/discussion/django_comment_client/utils.py
@@ -24,15 +24,7 @@ from lms.djangoapps.discussion.django_comment_client.permissions import (
     has_permission
 )
 from lms.djangoapps.discussion.django_comment_client.settings import MAX_COMMENT_DEPTH
-from openedx.core.djangoapps.course_groups.cohorts import get_cohort_id
-from openedx.core.djangoapps.discussions.utils import (
-    get_accessible_discussion_xblocks,
-    get_accessible_discussion_xblocks_by_course_id,
-    get_course_division_scheme,
-    get_discussion_categories_ids,
-    get_group_names_by_id,
-    has_required_keys,
-)
+from openedx.core.djangoapps.course_groups.cohorts import get_cohort_id, get_cohort_names, is_course_cohorted
 from openedx.core.djangoapps.django_comment_common.models import (
     FORUM_ROLE_COMMUNITY_TA,
     FORUM_ROLE_STUDENT,
@@ -122,6 +114,45 @@ def is_user_community_ta(user, course_id):
     Boolean operation to check whether a user's role is Community TA or not
     """
     return has_forum_access(user, course_id, FORUM_ROLE_COMMUNITY_TA)
+
+
+def has_required_keys(xblock):
+    """
+    Returns True iff xblock has the proper attributes for generating metadata
+    with get_discussion_id_map_entry()
+    """
+    for key in ('discussion_id', 'discussion_category', 'discussion_target'):
+        if getattr(xblock, key, None) is None:
+            log.debug(
+                "Required key '%s' not in discussion %s, leaving out of category map",
+                key,
+                xblock.location
+            )
+            return False
+    return True
+
+
+def get_accessible_discussion_xblocks(course, user, include_all=False):
+    """
+    Return a list of all valid discussion xblocks in this course that
+    are accessible to the given user.
+    """
+    include_all = getattr(user, 'is_community_ta', False)
+    return get_accessible_discussion_xblocks_by_course_id(course.id, user, include_all=include_all)
+
+
+@request_cached()
+def get_accessible_discussion_xblocks_by_course_id(course_id, user=None, include_all=False):  # pylint: disable=invalid-name
+    """
+    Return a list of all valid discussion xblocks in this course.
+    Checks for the given user's access if include_all is False.
+    """
+    all_xblocks = modulestore().get_items(course_id, qualifiers={'category': 'discussion'}, include_orphans=False)
+
+    return [
+        xblock for xblock in all_xblocks
+        if has_required_keys(xblock) and (include_all or has_access(user, 'load', xblock, course_id))
+    ]
 
 
 def get_discussion_id_map_entry(xblock):
@@ -432,6 +463,23 @@ def discussion_category_id_access(course, user, discussion_id, xblock=None):
         return has_required_keys(xblock) and (include_all or has_access(user, 'load', xblock, course.id))
     except DiscussionIdMapIsNotCached:
         return discussion_id in get_discussion_categories_ids(course, user)
+
+
+def get_discussion_categories_ids(course, user, include_all=False):
+    """
+    Returns a list of available ids of categories for the course that
+    are accessible to the given user.
+
+    Args:
+        course: Course for which to get the ids.
+        user:  User to check for access.
+        include_all (bool): If True, return all ids. Used by configuration views.
+
+    """
+    accessible_discussion_ids = [
+        xblock.discussion_id for xblock in get_accessible_discussion_xblocks(course, user, include_all=include_all)
+    ]
+    return course.top_level_discussion_topic_ids + accessible_discussion_ids
 
 
 class JsonResponse(HttpResponse):
@@ -833,7 +881,7 @@ def get_group_id_for_user(user, course_discussion_settings):
     If discussions are not divided, this method will return None.
     It will also return None if the user is in no group within the specified division_scheme.
     """
-    division_scheme = get_course_division_scheme(course_discussion_settings)
+    division_scheme = _get_course_division_scheme(course_discussion_settings)
     if division_scheme == CourseDiscussionSettings.COHORT:
         return get_cohort_id(user, course_discussion_settings.course_id)
     elif division_scheme == CourseDiscussionSettings.ENROLLMENT_TRACK:
@@ -912,7 +960,51 @@ def course_discussion_division_enabled(course_discussion_settings):
 
     Returns: True if discussion division is enabled for the course, else False
     """
-    return get_course_division_scheme(course_discussion_settings) != CourseDiscussionSettings.NONE
+    return _get_course_division_scheme(course_discussion_settings) != CourseDiscussionSettings.NONE
+
+
+def available_division_schemes(course_key):
+    """
+    Returns a list of possible discussion division schemes for this course.
+    This takes into account if cohorts are enabled and if there are multiple
+    enrollment tracks. If no schemes are available, returns an empty list.
+    Args:
+        course_key: CourseKey
+
+    Returns: list of possible division schemes (for example, CourseDiscussionSettings.COHORT)
+    """
+    available_schemes = []
+    if is_course_cohorted(course_key):
+        available_schemes.append(CourseDiscussionSettings.COHORT)
+    if enrollment_track_group_count(course_key) > 1:
+        available_schemes.append(CourseDiscussionSettings.ENROLLMENT_TRACK)
+    return available_schemes
+
+
+def enrollment_track_group_count(course_key):
+    """
+    Returns the count of possible enrollment track division schemes for this course.
+    Args:
+        course_key: CourseKey
+    Returns:
+        Count of enrollment track division scheme
+    """
+    return len(_get_enrollment_track_groups(course_key))
+
+
+def _get_course_division_scheme(course_discussion_settings):
+    division_scheme = course_discussion_settings.division_scheme
+    if (
+        division_scheme == CourseDiscussionSettings.COHORT and
+        not is_course_cohorted(course_discussion_settings.course_id)
+    ):
+        division_scheme = CourseDiscussionSettings.NONE
+    elif (
+        division_scheme == CourseDiscussionSettings.ENROLLMENT_TRACK and
+        enrollment_track_group_count(course_discussion_settings.course_id) <= 1
+    ):
+        division_scheme = CourseDiscussionSettings.NONE
+    return division_scheme
 
 
 def get_group_name(group_id, course_discussion_settings):
@@ -929,6 +1021,38 @@ def get_group_name(group_id, course_discussion_settings):
     """
     group_names_by_id = get_group_names_by_id(course_discussion_settings)
     return group_names_by_id[group_id] if group_id in group_names_by_id else None
+
+
+def get_group_names_by_id(course_discussion_settings):
+    """
+    Creates of a dict of group_id to learner-facing group names, for the division_scheme
+    in use as specified by course_discussion_settings.
+    Args:
+        course_discussion_settings: CourseDiscussionSettings model instance
+
+    Returns: dict of group_id to learner-facing group names. If no division_scheme
+    is in use, returns an empty dict.
+    """
+    division_scheme = _get_course_division_scheme(course_discussion_settings)
+    course_key = course_discussion_settings.course_id
+    if division_scheme == CourseDiscussionSettings.COHORT:
+        return get_cohort_names(get_course_by_id(course_key))
+    elif division_scheme == CourseDiscussionSettings.ENROLLMENT_TRACK:
+        # We negate the group_ids from dynamic partitions so that they will not conflict
+        # with cohort IDs (which are an auto-incrementing integer field, starting at 1).
+        return {-1 * group.id: group.name for group in _get_enrollment_track_groups(course_key)}
+    else:
+        return {}
+
+
+def _get_enrollment_track_groups(course_key):
+    """
+    Helper method that returns an array of the Groups in the EnrollmentTrackUserPartition for the given course.
+    If no such partition exists on the course, an empty array is returned.
+    """
+    partition_service = PartitionService(course_key)
+    partition = partition_service.get_user_partition(ENROLLMENT_TRACK_PARTITION_ID)
+    return partition.groups if partition else []
 
 
 def _verify_group_exists(group_id, course_discussion_settings):
