+++ lms/djangoapps/discussion/rest_api/api.py
@@ -19,7 +19,6 @@ from rest_framework.request import Request
 
 from lms.djangoapps.courseware.courses import get_course_with_access
 from lms.djangoapps.courseware.exceptions import CourseAccessRedirect
-from openedx.core.djangoapps.discussions.utils import get_accessible_discussion_xblocks
 from openedx.core.djangoapps.django_comment_common.comment_client.comment import Comment
 from openedx.core.djangoapps.django_comment_common.comment_client.thread import Thread
 from openedx.core.djangoapps.django_comment_common.comment_client.utils import CommentClientRequestError
@@ -63,6 +62,7 @@ from ..django_comment_client.base.views import (
     track_voted_event,
 )
 from ..django_comment_client.utils import (
+    get_accessible_discussion_xblocks,
     get_group_id_for_user,
     is_commentable_divided,
 )
