+++ lms/djangoapps/discussion/rest_api/views.py
@@ -16,10 +16,10 @@ from rest_framework.parsers import JSONParser
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework.viewsets import ViewSet
+from xmodule.modulestore.django import modulestore
 
 from lms.djangoapps.course_goals.models import UserActivity
 from lms.djangoapps.instructor.access import update_forum_role
-from openedx.core.djangoapps.discussions.serializers import DiscussionSettingsSerializer
 from openedx.core.djangoapps.django_comment_common import comment_client
 from openedx.core.djangoapps.django_comment_common.models import CourseDiscussionSettings, Role
 from openedx.core.djangoapps.user_api.accounts.permissions import CanReplaceUsername, CanRetireUser
@@ -27,7 +27,6 @@ from openedx.core.djangoapps.user_api.models import UserRetirementStatus
 from openedx.core.lib.api.authentication import BearerAuthenticationAllowInactiveUser
 from openedx.core.lib.api.parsers import MergePatchParser
 from openedx.core.lib.api.view_utils import DeveloperErrorViewMixin, view_auth_classes
-from xmodule.modulestore.django import modulestore
 from ..rest_api.api import (
     create_comment,
     create_thread,
@@ -52,6 +51,7 @@ from ..rest_api.forms import (
 from ..rest_api.serializers import (
     DiscussionRolesListSerializer,
     DiscussionRolesSerializer,
+    DiscussionSettingsSerializer,
 )
 
 log = logging.getLogger(__name__)
