+++ openedx/core/djangoapps/discussions/docs/decisions/0004-in-context-discussions-linking.rst
@@ -50,7 +50,7 @@ Consideration
 -------------
 
 Most of these configuration entries would be right at home in the
-``DiscussionsConfiguration`` model in ``plugin_settings``, however since they need
+`DiscussionsConfiguration` model in `plugin_settings`, however since they need
 to be available during course import-export, they should be stored in the
 course object itself.
 
@@ -59,13 +59,13 @@ using its usage key. This same mechanism can be used to associate a Unit usage
 key with a corresponding discussion id.
 
 However the current mechanism has a few issues. It is stored as a JSON
-structure in the ``DiscussionsIdMapping`` model which has course id and a mapping
+structure in the `DiscussionsIdMapping` model which has course id and a mapping
 of the discussion id to the xblock usage key in a single dict.
 
 This is OK for the existing setup because this is just a caching mechanism and
 the source of truth for this mapping is the XBlock itself, which stores the
 discussion id. On course publish this information is cached to
-``DiscussionsIdMapping``.
+`DiscussionsIdMapping`.
 
 For the new discussions system though, this mapping would be the source of
 truth for the link between discussions and units, so we should use a model
@@ -76,12 +76,12 @@ Decision
 
 Since the discussions settings need to be stored in the course structure we
 should create a new JSON structure in the course that matches the structure
-of ``plugin_settings``. This can then be used to store not just the settings
+of `plugin_settings`. This can then be used to store not just the settings
 for the inbuilt discussions provider, but for any discussions provider in the
 future.
 
-When a course is published, we can copy over all the ``plugin_settings`` to the
-course in a JSON field called ``discussions_settings`` with the following
+When a course is published, we can copy over all the `plugin_settings` to the
+course in a JSON field called `dicussions_settings` with the following
 structure:
 
 .. code-block:: JSON
@@ -95,7 +95,7 @@ structure:
         }
     }
 
-The ``edx-next`` key here represents the provider id, allowing for potentially
+The `edx-next` key here represents the provider id, allowing for potentially
 multiple provider configs to coexist in case of switching providers etc.
 Settings outside this key are those that are applicable to all providers. Note
 that they may not be supported by all providers though, in which case they will
@@ -103,16 +103,16 @@ simply be ignored.
 
 To store Unit-level discussions settings, we can simply add a boolean field
 to the Unit block that specifies whether it is discussable or not. To be
-consistent with the above names we can call this field ``discussions_enabled``.
+consistent with the above names we can call this field `discussions_enabled`.
 
 A signal can be created using the new Hooks extension system proposed in OEP-50
 that is triggered when discussions settings change. This signal can encapsulate
 all the data needed for setting up discussions from the modulestore. It can
 traverse through all teh Units in the course that match the criterion from the
-discussions settings and provide the needed details as part of the signal data.
+dicussions settings and provide the needed details as part of the signal data.
 
 A handler for the above signal, we create the discussion topics in
-``cs_comments_service`` and add a mapping. If an existing unit with discussions
+`cs_comments_service` and add a mapping. If an existing unit with discussions
 is removed, we will disable the link but not delete the data.
 
 The discussion grouping at subsections will simply combine the topics from all
