+++ openedx/core/djangoapps/discussions/serializers.py
@@ -1,16 +1,17 @@
 """
 Serializers for Discussion views.
 """
-from django.core.exceptions import ValidationError
+
 from lti_consumer.api import get_lti_pii_sharing_state_for_course
 from lti_consumer.models import LtiConfiguration
+from opaque_keys.edx.keys import CourseKey
 from rest_framework import serializers
+from xmodule.modulestore.django import modulestore
 
+from lms.djangoapps.discussion.rest_api.serializers import DiscussionSettingsSerializer
 from openedx.core.djangoapps.django_comment_common.models import CourseDiscussionSettings
 from openedx.core.lib.courses import get_course_by_id
-from xmodule.modulestore.django import modulestore
 from .models import AVAILABLE_PROVIDER_MAP, DEFAULT_PROVIDER_TYPE, DiscussionsConfiguration, Features
-from .utils import available_division_schemes, get_divided_discussions
 
 
 class LtiSerializer(serializers.ModelSerializer):
@@ -171,23 +172,10 @@ class DiscussionsConfigurationSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = DiscussionsConfiguration
-        course_fields = [
-            'provider_type',
-            'enable_in_context',
-            'enable_graded_units',
-            'unit_level_visibility',
-        ]
         fields = [
             'enabled',
-        ] + course_fields
-
-    def _get_course(self):
-        """
-        Get course and save it in the context, so it doesn't need to be reloaded.
-        """
-        if self.context.get('course') is None:
-            self.context['course'] = get_course_by_id(self.instance.context_key)
-        return self.context['course']
+            'provider_type',
+        ]
 
     def create(self, validated_data):
         """
@@ -199,11 +187,13 @@ class DiscussionsConfigurationSerializer(serializers.ModelSerializer):
         """
         Transform the *incoming* primitive data into a native value.
         """
-        payload = super().to_internal_value(data)
-        payload.update({
+        payload = {
+            'context_key': data.get('course_key', ''),
+            'enabled': data.get('enabled', False),
             'lti_configuration': data.get('lti_configuration', {}),
             'plugin_configuration': data.get('plugin_configuration', {}),
-        })
+            'provider_type': data.get('provider_type', DEFAULT_PROVIDER_TYPE),
+        }
         return payload
 
     def to_representation(self, instance: DiscussionsConfiguration) -> dict:
@@ -247,17 +237,14 @@ class DiscussionsConfigurationSerializer(serializers.ModelSerializer):
         """
         Update and save an existing instance
         """
-        # This needs to check which fields have changed, so do it before
-        # fields are copied over.
-        instance = self._update_course_configuration(instance, validated_data)
-        instance = self._update_plugin_configuration(instance, validated_data)
         for key in self.Meta.fields:
             value = validated_data.get(key)
             if value is not None:
                 setattr(instance, key, value)
         # _update_* helpers assume `enabled` and `provider_type`
         # have already been set
-        instance = self._update_lti(instance, validated_data)
+        instance = self._update_lti(instance, validated_data, instance.context_key)
+        instance = self._update_plugin_configuration(instance, validated_data)
         instance.save()
         return instance
 
@@ -265,6 +252,7 @@ class DiscussionsConfigurationSerializer(serializers.ModelSerializer):
         self,
         instance: DiscussionsConfiguration,
         validated_data: dict,
+        course_key: CourseKey
     ) -> DiscussionsConfiguration:
         """
         Update LtiConfiguration
@@ -280,7 +268,7 @@ class DiscussionsConfigurationSerializer(serializers.ModelSerializer):
                 data=lti_configuration_data,
                 partial=True,
                 context={
-                    'pii_sharing_allowed': get_lti_pii_sharing_state_for_course(instance.context_key),
+                    'pii_sharing_allowed': get_lti_pii_sharing_state_for_course(course_key),
                 }
             )
             if lti_serializer.is_valid(raise_exception=True):
@@ -296,140 +284,23 @@ class DiscussionsConfigurationSerializer(serializers.ModelSerializer):
         """
         Create/update legacy provider settings
         """
-        plugin_configuration = validated_data.pop('plugin_configuration', {})
         updated_provider_type = validated_data.get('provider_type') or instance.provider_type
         will_support_legacy = bool(
             updated_provider_type == 'legacy'
         )
         if will_support_legacy:
+            course_key = instance.context_key
+            course = get_course_by_id(course_key)
             legacy_settings = LegacySettingsSerializer(
-                self._get_course(),
+                course,
                 context={
                     'user_id': self.context['user_id'],
                 },
-                data=plugin_configuration,
+                data=validated_data.get('plugin_configuration', {}),
             )
             if legacy_settings.is_valid(raise_exception=True):
                 legacy_settings.save()
-            instance.plugin_configuration = {
-                "group_at_subsection": plugin_configuration.get("group_at_subsection", False)
-            }
+            instance.plugin_configuration = {}
         else:
-            instance.plugin_configuration = plugin_configuration
-        return instance
-
-    def _update_course_configuration(
-        self,
-        instance: DiscussionsConfiguration,
-        validated_data: dict,
-    ) -> DiscussionsConfiguration:
-        """
-        Update configuration settings that are stored in the course.
-        """
-        save = False
-        updated_provider_type = validated_data.get('provider_type') or instance.provider_type
-        for key in self.Meta.course_fields:
-            value = validated_data.get(key)
-            # Delay loading course till we know something has actually been updated
-            if value is not None and value != getattr(instance, key):
-                self._get_course().discussions_settings[key] = value
-                save = True
-        new_plugin_config = validated_data.get('plugin_configuration', None)
-        if new_plugin_config and new_plugin_config != instance.plugin_configuration:
-            save = True
-            # Any fields here that aren't already stored in the course structure
-            # or in other models should be stored here.
-            self._get_course().discussions_settings[updated_provider_type] = {
-                key: value
-                for key, value in new_plugin_config.items()
-                if (
-                    key not in LegacySettingsSerializer.Meta.fields and
-                    key not in LegacySettingsSerializer.Meta.fields_cohorts
-                )
-            }
-        if save:
-            modulestore().update_item(self._get_course(), self.context['user_id'])
-        return instance
-
-
-class DiscussionSettingsSerializer(serializers.Serializer):
-    """
-    Serializer for course discussion settings.
-    """
-    divided_discussions = serializers.ListField(
-        child=serializers.CharField(),
-        write_only=True,
-    )
-    divided_course_wide_discussions = serializers.ListField(
-        child=serializers.CharField(),
-        read_only=True,
-    )
-    divided_inline_discussions = serializers.ListField(
-        child=serializers.CharField(),
-        read_only=True,
-    )
-    always_divide_inline_discussions = serializers.BooleanField()
-    division_scheme = serializers.CharField()
-
-    def to_internal_value(self, data: dict) -> dict:
-        """
-        Transform the *incoming* primitive data into a native value.
-        """
-        payload = super().to_internal_value(data) or {}
-        course = self.context['course']
-        instance = self.context['settings']
-        if any(item in data for item in ('divided_course_wide_discussions', 'divided_inline_discussions')):
-            divided_course_wide_discussions, divided_inline_discussions = get_divided_discussions(
-                course, instance
-            )
-            divided_course_wide_discussions = data.get(
-                'divided_course_wide_discussions',
-                divided_course_wide_discussions
-            )
-            divided_inline_discussions = data.get('divided_inline_discussions', divided_inline_discussions)
-            try:
-                payload['divided_discussions'] = divided_course_wide_discussions + divided_inline_discussions
-            except TypeError as error:
-                raise ValidationError(str(error)) from error
-        for item in ('always_divide_inline_discussions', 'division_scheme'):
-            if item in data:
-                payload[item] = data[item]
-        return payload
-
-    def to_representation(self, instance: CourseDiscussionSettings) -> dict:
-        """
-        Return a serialized representation of the course discussion settings.
-        """
-        payload = super().to_representation(instance)
-        course = self.context['course']
-        instance = self.context['settings']
-        course_key = course.id
-        divided_course_wide_discussions, divided_inline_discussions = get_divided_discussions(
-            course, instance
-        )
-        payload = {
-            'id': instance.id,
-            'divided_inline_discussions': divided_inline_discussions,
-            'divided_course_wide_discussions': divided_course_wide_discussions,
-            'always_divide_inline_discussions': instance.always_divide_inline_discussions,
-            'division_scheme': instance.division_scheme,
-            'available_division_schemes': available_division_schemes(course_key)
-        }
-        return payload
-
-    def create(self, validated_data):
-        """
-        This method intentionally left empty
-        """
-
-    def update(self, instance: CourseDiscussionSettings, validated_data: dict) -> CourseDiscussionSettings:
-        """
-        Update and save an existing instance
-        """
-        if not any(field in validated_data for field in self.fields):
-            raise ValidationError('Bad request')
-        try:
-            instance.update(validated_data)
-        except ValueError as e:
-            raise ValidationError(str(e)) from e
+            instance.plugin_configuration = validated_data.get('plugin_configuration') or {}
         return instance
