+++ openedx/core/djangoapps/discussions/tests/test_models.py
@@ -162,7 +162,7 @@ class DiscussionsConfigurationModelTest(TestCase):
         assert configuration.enabled  # by default
         assert configuration.lti_configuration is None
         assert len(configuration.plugin_configuration.keys()) == 0
-        assert configuration.provider_type == 'legacy'
+        assert not configuration.provider_type
 
     def test_get_with_values(self):
         """
@@ -236,7 +236,7 @@ class DiscussionsConfigurationModelTest(TestCase):
         assert configuration.enabled
         assert not configuration.lti_configuration
         assert not configuration.plugin_configuration
-        assert configuration.provider_type == 'legacy'
+        assert not configuration.provider_type
 
     def test_get_explicit(self):
         """
